+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Check that buildDir exists before loading tests
 - Add CTest path autodetect from CMakeCache.txt file; this should avoid useless creation of `Testing` subdirs in non-CMake project dirs
 
+### Changed
+
+- Add better error message when using CTest version < 3.14
+
 ## [0.1.0] - 2019-08-19
 
 ### Added